Moved WebSocketMessage to core Wolverine, added message type name rule for web sockets
